Fix cmd_exists() for both positive and negative cases
The previous use of subprocess.call() would always return true, even for
commands that do not exist.

Since this function does not see non-sanitized input for the pinentry
path, shell=False is not absolutely necessary; however if it does need
to be added, "which" should be used instead of "type" as on some
systems, "type" is a shell built-in.

- Added negative test for cmd_exists() in addition to the older positive
test (now renamed)